# business-central-backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Business Central Backend API
  version: 0.1.0
  description: |
    Contract-first OpenAPI documentation for the Business Central backend.

    Health, authentication, authorization, commerce, reporting, service, and
    repair endpoints are implemented. Reusable schemas, security conventions,
    response envelopes, required error codes, and collection query semantics
    are documented here.
  license:
    name: Proprietary
  x-business-central-status: implemented-current-scope
servers:
  - url: http://localhost:8080
    description: Local backend (when the HTTP runtime is available)
...
paths:
...
  /api/v1/admin/currencies:
    get:
      tags: [Merchants]
      summary: List platform currencies as an administrator
      operationId: adminListCurrencies
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Platform currencies.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrencyCollectionEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Merchants]
      summary: Create a platform currency
      operationId: createCurrency
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCurrencyRequest'
      responses:
        '201':
          description: Platform currency created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrencyEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/admin/currencies/{code}:
    parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          minLength: 3
          maxLength: 3
          pattern: '^[A-Za-z]{3}$'
    get:
      tags: [Merchants]
      summary: Get a platform currency
      operationId: getCurrency
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Platform currency.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrencyEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Merchants]
      summary: Update a platform currency
      operationId: updateCurrency
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCurrencyRequest'
      responses:
        '200':
          description: Platform currency updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrencyEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Merchants]
      summary: Delete a platform currency
      operationId: deleteCurrency
      security: [{ bearerAuth: [] }]
      responses:
        '204':
          description: Platform currency deleted.
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /api/v1/auth/refresh:
    post:
      tags: [Authentication]
      summary: Rotate a refresh token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: A new access and refresh token were issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/me:
    get:
      tags: [Authentication]
      summary: Get the authenticated user
      operationId: currentUser
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: The authenticated user and active membership.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/logout:
    post:
      tags: [Authentication]
      summary: Revoke the current refresh token
      operationId: logout
      security: [{ bearerAuth: [] }]
      responses:
        '204':
          description: The refresh token was revoked.
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/admin/merchants:
    get:
      tags: [Merchants]
      summary: List merchants
      operationId: listMerchants
      security: [{ bearerAuth: [] }]
      description: Only a platform-admin identity may call this operation.
      parameters:
        - $ref: '#/components/parameters/Query'
        - $ref: '#/components/parameters/Filter'
        - $ref: '#/components/parameters/PageIndex'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: A paginated list of merchants.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantCollectionEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Merchants]
      summary: Create a merchant and default roles
      operationId: createMerchantAccount
      security: [{ bearerAuth: [] }]
      description: Only a platform-admin identity may call this operation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMerchantAccountRequest'
      responses:
        '201':
          description: The merchant and default Manager and Staff roles were created. No user account is created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantAccountEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/admin/merchant-users:
    post:
      tags: [Merchants, Users]
      summary: Create a merchant and its owner user
      operationId: createMerchantUser
      security: [{ bearerAuth: [] }]
      description: >-
        Only a platform-admin identity may call this operation. It atomically
        creates the merchant, system roles, owner user, membership, and owner
        role assignment.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMerchantUserRequest'
      responses:
        '201':
          description: The merchant and owner user were created together.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantUserProvisioningEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/admin/merchants/{id}:
    patch:
      tags: [Merchants]
      summary: Update merchant settings or active state
      operationId: updateMerchant
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMerchantRequest'
      responses:
        '200':
          description: The merchant was updated. Setting is_active false deactivates access without deleting the merchant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/admin/merchants/{id}/roles:
    get:
      tags: [Users]
      summary: List roles for a merchant
      operationId: listMerchantRoles
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Roles available for user membership assignment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleCollectionEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Users]
      summary: Create a role for a merchant
      operationId: createMerchantRole
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
      responses:
        '201':
          description: Merchant role created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/admin/merchants/{id}/roles/{roleId}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      - name: roleId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
    get:
      tags: [Users]
      summary: Get a merchant role
      operationId: getMerchantRole
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Merchant role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Users]
      summary: Update a merchant role
      operationId: updateMerchantRole
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
      responses:
        '200':
          description: Merchant role updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Users]
      summary: Delete a custom merchant role
      operationId: deleteMerchantRole
      description: System roles cannot be deleted. Removing a custom role also removes its membership assignments.
      security: [{ bearerAuth: [] }]
      responses:
        '204':
          description: Merchant role deleted.
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/admin/permissions:
    get:
      tags: [Users]
      summary: List permissions available for role assignment
      operationId: listPermissions
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Platform permission catalogue.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionCollectionEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v1/users/{id}:
    parameters:
      - $ref: '#/components/parameters/UserID'
      - $ref: '#/components/parameters/MerchantIDHeader'
    get:
      tags: [Users]
      summary: Get a user membership
      operationId: getUser
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: The requested user membership.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Users]
      summary: Update a user membership
      operationId: updateUser
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: The user membership was updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Users]
      summary: Deactivate a user membership
      operationId: deleteUser
      security: [{ bearerAuth: [] }]
      responses:
        '204':
          description: The membership was deactivated.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Send the access token returned by login or refresh as
        Authorization: Bearer <access-token>. Refresh tokens are rotated by
        POST /api/v1/auth/refresh.
...
  responses:
...
    Forbidden:
      description: The authenticated principal lacks the required permission.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: The command conflicts with the current resource state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    CreateCurrencyRequest:
      type: object
      required: [code, name, decimal_places]
      properties:
        code:
          type: string
          minLength: 3
          maxLength: 3
          pattern: '^[A-Za-z]{3}$'
        name:
          type: string
          minLength: 1
        symbol:
          type: string
          nullable: true
        decimal_places:
          type: integer
          minimum: 0
          maximum: 6
...
    CurrencyEnvelope:
      allOf: [{ $ref: '#/components/schemas/ResourceEnvelope' }]
    CurrencyCollectionEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Currency'
        meta:
          $ref: '#/components/schemas/ResourceMeta'
...